feat(views): moved controllers to views;
- Moved majority of controller classes and methods to views.
- Added `is_registry` helper.
- Minor template fixes.
- Removed `IRoutes` implementations from plugin classes.
