+++ ckanext/canada/helpers.py
@@ -38,6 +38,10 @@ GEO_MAP_TYPE_OPTION = 'wet_theme.geo_map_type'
 GEO_MAP_TYPE_DEFAULT = 'static'
 
 
+def is_registry():
+    # type: () -> bool
+    return 'canada_internal' in config.get('ckan.plugins', [])
+
 
 def get_translated_t(data_dict, field):
     '''
@@ -146,7 +150,7 @@ def catalogue_last_update_date():
 
 def today():
     return datetime.datetime.now(EST()).strftime("%Y-%m-%d")
-    
+
 # Return the Date format that the WET datepicker requires to function properly
 def date_format(date_string):
     if not date_string:
@@ -199,7 +203,7 @@ def remove_duplicates(a_list):
     s = set()
     for i in a_list:
         s.add(i)
-            
+
     return s
 
 def get_license(license_id):
@@ -239,7 +243,7 @@ def adobe_analytics_lang():
 
 def adobe_analytics_js():
     return str(config.get('adobe_analytics.js', ''))
-    
+
 def loop11_key():
     return str(config.get('loop11.key', ''))
 
@@ -247,14 +251,14 @@ def drupal_session_present(request):
     for name in request.cookies.keys():
         if name.startswith("SESS"):
             return True
-    
+
     return False
-    
+
 def parse_release_date_facet(facet_results):
     counts = facet_results['counts'][1::2]
     ranges = facet_results['counts'][0::2]
     facet_dict = dict()
-    
+
     if len(counts) == 0:
         return dict()
     elif len(counts) == 1:
@@ -263,9 +267,9 @@ def parse_release_date_facet(facet_results):
         else:
             facet_dict = {'scheduled': {'count': counts[0], 'url_param': '[' + ranges[0] + ' TO ' + facet_results['end'] + ']'} }
     else:
-        facet_dict = {'published': {'count': counts[0], 'url_param': '[' + ranges[0] + ' TO ' + ranges[1] + ']'} , 
+        facet_dict = {'published': {'count': counts[0], 'url_param': '[' + ranges[0] + ' TO ' + ranges[1] + ']'} ,
                       'scheduled': {'count': counts[1], 'url_param': '[' + ranges[1] + ' TO ' + facet_results['end'] + ']'} }
-    
+
     return facet_dict
 
 def is_ready_to_publish(package):
