+++ ckanext/canada/templates/internal/admin/base.html
@@ -7,6 +7,6 @@
 {% block content_primary_nav %}
   {{ h.build_nav_icon('user.index', _('Users'), icon='user') }}
   {{ h.build_nav_icon('admin.trash', _('Trash'), icon='trash-o') }}
-  {{ h.build_nav_icon('ckanadmin_publish', _('Publish Records')) }}
+  {{ h.build_nav_icon('canada.ckanadmin_publish', _('Publish Records'), icon='cloud-upload') }}
   {{ h.build_extra_admin_nav() }}
 {% endblock %}
