+++ ckanext/canada/templates/internal/admin/snippets/publish_package_list.html
@@ -1,19 +1,19 @@
 {% set list_class = 'list-bullet-none' %}
 {% if packages %}
-<form action="{{ h.url_for(controller='ckanext.canada.controller:CanadaAdminController', action='publish') }}" method="POST" id="publish_form">
-  
+<form action="{{ h.url_for('canada.ckanadmin_publish_datasets') }}" method="POST" id="publish_form">
+
   <div>
     <label for="publish_all" class="form-checkbox" style="float:right; display:inline-block; margin-top:-25px;">
       <input type="checkbox" id="publish_all" name="publish_all">{{ _('Select All for Publishing') }}
     </label>
   </div>
-  
+
   <ul class="list-unstyled">
     {% for package in packages %}
       {% snippet 'admin/snippets/package_item.html', package=package, item_class=item_class, hide_resources=hide_resources, banner=banner, truncate=truncate, truncate_title=truncate_title %}
     {% endfor %}
   </ul>
-  
+
   <fieldset>
     <div>
       <label for="publish_date">
@@ -25,9 +25,9 @@
    	  </label>
       <input type="date" id="publish_date" name="publish_date" min="{{h.today()}}" value="{{h.today()}}" />
     </div>
-    
+
     <button type="submit" class="button button-large button-attention">{{ _('Publish Selected Records') }}</button>
-    
+
   </fieldset>
 </form>
 {% endif %}
