+++ ckanext/canada/templates/internal/xloader/resource_data.html
@@ -3,7 +3,7 @@
 {% block primary_content_inner %}
   {% if res.datastore_active %}
     <form>  {# form required for button styling :-| #}
-      <a href="{% url_for controller='ckanext.canada.controller:CanadaDatastoreController', action='delete_datastore_table', id=pkg.id, resource_id=res.id %}"
+      <a href="{% url_for 'canada.delete_datastore_table', id=pkg.id, resource_id=res.id %}"
         class="btn btn-danger pull-left"
         data-module="confirm-action" data-module-content="{{ _('Are you sure you want to delete this DataStore table and Data Dictionary?') }}"
         >{% block delete_datastore_button_text %}<i class="fa fa-remove"></i>{{ _('Delete DataStore table') }}{% endblock %}</a>
